fix(deploy): rename src/data to src/data_layer and update imports to resolve ModuleNotFoundError on GCP
Also bumped version to 3.1.3
