+++ pyproject.toml
@@ -3,7 +3,7 @@
 # =============================================================================
 [project]
 name = "crypto-auto-trading-bot"
-version = "3.1"
+version = "3.1.3"
 description = "Automated cryptocurrency trading bot supporting multiple exchanges"
 readme = "README.md"
 license = {text = "MIT"}
