+++ src/__init__.py
@@ -1,2 +1,2 @@
 # this file can be empty
-VERSION = "3.1.2"
+VERSION = "3.1.3"
