+++ src/data_layer/data_manager.py
@@ -8,9 +8,9 @@ from typing import Any
 import pandas as pd
 
 from src.core.models.index import IndexType
-from src.data.data_collector import DataCollector
-from src.data.data_processor import DataProcessor
-from src.data.data_saver import DataSaver
+from src.data_layer.data_collector import DataCollector
+from src.data_layer.data_processor import DataProcessor
+from src.data_layer.data_saver import DataSaver
 from src.infrastructure.logging.set_logger import get_adapter, get_logger
 from src.interfaces.pipeline_interface import PipelineController
